+++ src/main/java/racinggame/RacingGameApplication.java
@@ -1,9 +1,16 @@
 package racinggame;
 
+import racinggame.controller.RacingGame;
+import racinggame.component.RandomNumberGenerator;
+import racinggame.domain.Car;
+import racinggame.model.RacingGameResult;
+import racinggame.view.InputView;
+import racinggame.view.OutputView;
+
 import java.util.List;
 import java.util.stream.Collectors;
 
-public class Main {
+public class RacingGameApplication {
 
     public static void main(String[] args) {
         List<Car> cars = InputView.inputCarNames().stream()
