+++ src/main/java/racinggame/common/StringConstants.java
@@ -1,4 +1,4 @@
-package racinggame;
+package racinggame.common;
 
 public class StringConstants {
