+++ src/test/java/racinggame/domain/CarTest.java
@@ -1,7 +1,8 @@
-package racinggame;
+package racinggame.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racinggame.domain.Car;
 
 import static org.assertj.core.api.Assertions.assertThat;
